Handle non active payment intents (#2132)
Currently when retrieving payment intents to use in Stripe, we aren't
considering whether they have been canceled or are already succeeded.
This PR looks to add an `active` scope to the cart payment intents
relationship and also, when an intent is canceled, we now set the status
on the relevant model.