Dont split name when storing Stripe address from Webhook (#2147)
Currently when storing the address that comes back from Stripe, we are
trying to split the name and populating the result in the first/last
name columns. This creates issues when the name Stripe returns contains
more than just the first and last name.

Therefore it would be more benefical to just simple store the name in
the `first_name` column to prevent any loss of data since there
shouldn't be any side effects of doing this.