Allow Stripe Webhook parameter fetching to be customised. (#2173)
Currently the Stripe Webhook needs to find a PaymentIntent ID and an
Order ID so that the order can be found and processed.

At the moment this is hardcoded to assume the PaymentIntent might have
an `order_id` value present in the metadata. This is a bit of an
assumption and although this works with the Stripe package, since it
adds that value, it would still be useful for developers to be able to
customise this behaviour, in the event they need to get the draft order
ID from somewhere else.

This PR puts the logic for retrieving the intent ID and draft order ID
into an action which can be swapped out for a different class, if the
developer needs this.

---------

Co-authored-by: Author <actions@github.com>
Co-authored-by: Glenn Jacobs <glenn@neondigital.co.uk>